Add TranslationContext to ShelleyLedgerConfig
Because node will need access to TranslationContext while
transitioning between eras we must make that TranslationContext
available to the node. We will do it by making it available in in
ShelleyLedgerConfig.

This changed required modification to how NoThunk is derived for
ShelleyLedgerConfig which ultimately led to modification of
ShelleyBasedEra typeclass (requiring TranslationContext era for
ShelleyBasedEra era instance).