Change Alonzo's TranslationContext to Int
Alozno's TranslationContext should (and will) be provided by
ledger-spec. For now we provide that instance in the
Ouroboros/Consensus/Shelley/Orphans.hs.

Prior to this commit that instance was set to Unit (as it was for
other previous eras). We don't know yet what type it will be, but what
we do know is the fact that it is not going to be Unit. Thus this
commit is changing the instance to be Int and it's making required
changes to the code base for it to compile.